requirements.txt
attr==0.3.2
bpython==0.25
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.7
cx_Oracle==8.3.0
Cython==3.0.8
django_environ==0.12.0
djangorestframework==3.16.1
dl==0.1.0
docutils==0.22
filelock==3.19.1
HTMLParser==0.0.2
ipdb==0.13.13
ipython==8.12.3
ipywidgets==8.1.1
isort==6.0.1
jnius==1.1.0
keyring==25.6.0
MySQL-python==1.2.5
numpy==2.3.3
oracledb==3.3.0
ordereddict==1.1
Pillow==11.3.0
protobuf==6.32.0
psycopg==3.2.9
psycopg2==2.9.10
psycopg_pool==3.2.6
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==25.1.0
pytest==8.0.0
pywatchman==3.0.0
pywin32==306
PyYAML==6.0.1
PyYAML==6.0.2
Sphinx==8.2.3
tblib==3.1.0
thread==2.0.5
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
